tests/wpt/tests/layout-instability/outline.html

<!DOCTYPE html>
<title>Layout Instability: outline doesn't contribute to layout shift</title>
<link rel="help" href="https://wicg.github.io/layout-instability/" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/util.js"></script>
<div id="target" style="width: 300px; height: 300px; background: blue"></div>
<script>
promise_test(async () => {
  const watcher = new ScoreWatcher;

  // Wait for the initial render to complete.
  await waitForAnimationFrames(2);

  // Add outline for target. This should not generate a shift.
  target.style.outline = "10px solid blue";

  await waitForAnimationFrames(3);
  assert_equals(watcher.score, 0);
}, "Outline.");
</script>
